Add a "service" argument to system  and request specs
This allows us to point system and request specs to a specific service's host
